Extract TaskTokenizer to it's own class
