+++ src/main/java/seedu/toluist/controller/TaskController.java
@@ -1,15 +1,10 @@
 package seedu.toluist.controller;
 
-import java.util.HashMap;
 import java.util.logging.Logger;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import seedu.toluist.commons.core.LogsCenter;
 import seedu.toluist.dispatcher.CommandResult;
-import seedu.toluist.model.Task;
 import seedu.toluist.ui.Ui;
-import seedu.toluist.ui.UiStore;
 
 /**
  * TaskController is responsible for management of task
@@ -17,14 +12,6 @@ import seedu.toluist.ui.UiStore;
 public abstract class TaskController extends Controller {
 
     private String commandTemplate;
-
-    protected static final String TASK_VIEW_INDEX = "index";
-    protected static final String TASK_DESCRIPTION = "description";
-    protected static final String TASK_START_DATE_KEYWORD = "startdate/";
-    protected static final String TASK_END_DATE_KEYWORD = "enddate/";
-
-    protected static final int START_INDEX = 0;
-    protected static final int INVALID_INDEX = -1;
 
     protected Logger logger = LogsCenter.getLogger(getClass());
 
@@ -35,45 +22,6 @@ public abstract class TaskController extends Controller {
     public TaskController(Ui renderer, String commandTemplate) {
         super(renderer);
         this.commandTemplate = commandTemplate;
-    }
-
-    public HashMap<String, String> tokenize(String commandArgs, boolean hasIndex, boolean hasDescription) {
-        Pattern pattern = Pattern.compile(commandTemplate);
-        Matcher matcher = pattern.matcher(commandArgs.trim());
-        matcher.find();
-        HashMap<String, String> tokens = new HashMap<>();
-        if (hasIndex) {
-            tokens.put(TASK_VIEW_INDEX, matcher.group(TASK_VIEW_INDEX));
-        }
-        if (hasDescription) {
-            String description = matcher.group(TASK_DESCRIPTION);
-            int startDateIndex = description.lastIndexOf(TASK_START_DATE_KEYWORD);
-            int endDateIndex = description.lastIndexOf(TASK_END_DATE_KEYWORD);
-            String startDate = null;
-            String endDate = null;
-            if (endDateIndex != INVALID_INDEX) {
-                // This is a task with deadline
-                endDate = description.substring(endDateIndex + TASK_END_DATE_KEYWORD.length());
-                description = description.substring(START_INDEX, endDateIndex);
-                if (startDateIndex != INVALID_INDEX) {
-                    // This is an event
-                    startDate = description.substring(startDateIndex + TASK_START_DATE_KEYWORD.length());
-                    description = description.substring(START_INDEX, startDateIndex);
-                }
-            } // Else this is a floating task
-            // Note: We are not dealing with tasks that have only a start date.
-
-            tokens.put(TASK_DESCRIPTION, description);
-            tokens.put(TASK_START_DATE_KEYWORD, startDate);
-            tokens.put(TASK_END_DATE_KEYWORD, endDate);
-        }
-        return tokens;
-    }
-
-    public Task getTask(String indexToken) {
-        int index = indexToken == null ? INVALID_INDEX : Integer.parseInt(indexToken) - 1;
-        Task task = indexToken == null ? null : UiStore.getInstance().getTasks().get(index);
-        return task;
     }
 
     @Override
